requirements.txt
aiogram==2.25.1
aiohttp==3.8.6
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
djangorestframework==3.14.0
drf-spectacular==0.26.5
frozenlist==1.4.1
idna==3.6
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
magic-filter==1.0.12
multidict==6.0.4
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
rpds-py==0.15.2
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
yarl==1.9.4
